# nextjs/public/openapi.yaml
openapi: 3.1.0
info:
  title: VidTempla API
  version: "1.0"
  description: |
    VidTempla is an API-first platform for managing YouTube video descriptions with templates, containers, and variables.
    The REST API lets AI agents securely manage YouTube channels without handling OAuth directly.

    ## Authentication
    All requests require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer vtk_your_key_here
    ```
    API keys have one of two permission levels:
    - **read** — can only fetch data (GET requests and POST /analytics/query)
    - **read-write** — can also create, update, and delete resources

    Generate keys from the dashboard at Settings > API Keys.

    ## Response Envelope
    Every response uses a standard envelope:
    ```json
    { "data": { ... }, "error": null, "meta": { ... } }
    ```

    ## Pagination
    All list endpoints use cursor-based pagination. Default limit is 50, max is 100.
    ```
    GET /videos?cursor=...&limit=50
    ```
  contact:
    name: VidTempla
    url: https://vidtempla.com
  license:
    name: Proprietary
servers:
  - url: https://vidtempla.com/api/v1
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key starting with vtk_. Generate from Settings > API Keys."
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          type: "null"
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            suggestion:
              type: string
              description: Actionable suggestion for agents to self-correct
            status:
              type: integer
              description: HTTP status code
            meta:
              type: object
              additionalProperties: true
              description: Optional machine-readable context about the error
          required: [code, message, status]
      required: [data, error]
    PaginationMeta:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page. Null on last page.
        hasMore:
          type: boolean
          description: Whether more results exist beyond this page
        total:
          type: integer
          description: Total count of matching records
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INVALID_API_KEY
              message: API key not found or has been revoked
              suggestion: Generate a new key from Settings > API Keys
              status: 401
    InsufficientPermissions:
      description: API key lacks write permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INSUFFICIENT_PERMISSIONS
              message: This API key has read-only access
              suggestion: Create a new API key with read-write permission from Settings > API Keys
              status: 403
    QuotaExceeded:
      description: YouTube API quota or rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: QUOTA_EXCEEDED
              message: YouTube API quota exhausted
              suggestion: Retry with backoff. If daily quota is exhausted, wait for the Pacific midnight reset or contact support.
              status: 429
              meta:
                upstreamStatus: 403
                reasons: [quotaExceeded]
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
              suggestion: Try again later
              status: 500
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /youtube/comments:
    get:
      operationId: searchChannelComments
      summary: Search a channel's comments
      description: |
        Searches every comment thread related to a channel, narrowed by `searchTerms`
        when given. This is the discovery step of a link sweep: search for the old URL,
        review the matches, then send the comment IDs to `POST /youtube/comments/bulk-update`.

        VidTempla stores no comment state, so every call re-reads YouTube live.
      tags: [YouTube Comments]
      x-quota-cost: 1
      parameters:
        - name: channelId
          in: query
          required: true
          schema:
            type: string
          description: YouTube channel ID (UC...) of a connected channel
        - name: searchTerms
          in: query
          schema:
            type: string
          description: Return only threads whose text matches these terms, such as the URL being replaced
        - name: maxResults
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Max results per page (max 100)
        - name: order
          in: query
          schema:
            type: string
            enum: [relevance, time]
            default: relevance
          description: Sort order (relevance for top comments, time for newest)
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination token from the previous page's meta.cursor
      responses:
        "200":
          description: Matching comment threads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      description: YouTube commentThread resource
                  error:
                    type: "null"
                  meta:
                    type: object
                    properties:
                      cursor:
                        type: string
                        nullable: true
                        description: Pass as ?cursor= for the next page. Null on the last page.
                      hasMore:
                        type: boolean
                      total:
                        type: integer
                        nullable: true
                        description: Always null. YouTube reports no total for a comment search.
                      quotaUnits:
                        type: integer
                        example: 1
        "400":
          description: Missing channelId, invalid parameters, or a channelId that is not a UC... ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/QuotaExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /youtube/comments/reply:
    post:
      operationId: replyToComment
      summary: Reply to a comment
      description: |
        Posts a reply to an existing YouTube comment. A reply is new content, so no
        edit-log row is written — only updates and deletes are snapshotted.
      tags: [YouTube Comments]
      x-quota-cost: 50
      x-requires-write: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId, parentId, text]
              properties:
                channelId:
                  type: string
                  description: YouTube channel ID
                parentId:
                  type: string
                  description: The comment ID to reply to
                text:
                  type: string
                  description: Reply text
            example:
              channelId: "UCxxxxxxxxxxxxxxxxxxxxxx"
              parentId: "UgxAbCdEfGhIjKlMnOp"
              text: "Thanks for watching! Glad you enjoyed it."
      responses:
        "200":
          description: Reply posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: YouTube comment resource
                  error:
                    type: "null"
                  meta:
                    type: object
                    properties:
                      quotaUnits:
                        type: integer
                        example: 50
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InsufficientPermissions"
        "429":
          $ref: "#/components/responses/QuotaExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /youtube/comments/edits:
    get:
      operationId: listCommentEdits
      summary: List the comment edit log
      description: |
        Returns every comment update and deletion made through VidTempla for this
        workspace, newest first. Each row holds the comment's text as it was before the
        write — the only surviving copy, because YouTube keeps no comment history.

        Only rows with `textSource: "original"` hold restorable text. A `"display"` row
        is the HTML-marked-up rendering of a third-party comment and is an audit record
        only. Restoring is a normal write at 51 units per comment, not a free rollback.

        `status` is `pending` while a write is in flight, then `applied` or `failed`.
        A row stuck at `pending` is resolved by the next bulk update for the same channel;
        `unknown` means that reconciliation could not tell what happened.

        No YouTube call, so no quota is consumed.
      tags: [YouTube Comments]
      x-quota-cost: 0
      parameters:
        - name: channelId
          in: query
          schema:
            type: string
          description: Only return rows written with this channel's token
        - name: commentId
          in: query
          schema:
            type: string
          description: Only return rows for this comment
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque cursor from the previous page's meta.cursor
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Rows per page (max 100)
      responses:
        "200":
          description: Edit-log rows, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        channelId:
                          type: string
                        commentId:
                          type: string
                        videoId:
                          type: string
                          nullable: true
                        verb:
                          type: string
                          enum: [update, delete]
                        textSource:
                          type: string
                          enum: [original, display]
                          description: original is restorable; display is an audit record only
                        beforeText:
                          type: string
                        afterText:
                          type: string
                          nullable: true
                          description: Null for a delete
                        status:
                          type: string
                          enum: [pending, applied, failed, unknown]
                        source:
                          type: string
                          enum: [mcp, rest, dashboard]
                        createdAt:
                          type: string
                          format: date-time
                  error:
                    type: "null"
                  meta:
                    allOf:
                      - $ref: "#/components/schemas/PaginationMeta"
                      - type: object
                        properties:
                          quotaUnits:
                            type: integer
                            example: 0
        "400":
          description: Invalid cursor or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── YouTube Thumbnails ─────────────────────────────────────
...